ceph.mon.default: use service.running
Instead of starting and enabling the service ourselves, farm that job out to
service.running like we already do for the MGRs.

If the service starts and then fails almost immediately thereafter (as Ceph
daemons are known to do sometimes), DeepSea will merrily continue on its
way. Since it's better to fail where the failure actually arises, add a small
bash script that waits a short time and checks the service status, so we don't
continue with a failed MON.

Signed-off-by: Nathan Cutler <ncutler@suse.com>